+++ source/egov-erp/egov-eis/src/main/java/org/egov/eis/repository/AssignmentRepository.java
@@ -29,9 +29,9 @@ public interface AssignmentRepository extends JpaRepository<Assignment,Long>{
 	public List<Assignment> getAllActiveAssignmentsByEmpId(@Param("idPersonalInformation")Integer idPersonalInformation);
 	
 	@Query(" from Assignment A where A.fromDate<=:givenDate and A.toDate>=:givenDate and A.position.id=:posId order by A.fromDate")
-	public List<Assignment> getAssignmentsForPosition(@Param("posId")Integer posId,@Param("givenDate")Date givenDate);
+	public List<Assignment> getAssignmentsForPosition(@Param("posId")Long posId,@Param("givenDate")Date givenDate);
 	
 	@Query(" from Assignment A where A.fromDate<=current_date and A.toDate>=current_date and A.position.id=:posId")
-	public Assignment getPrimaryAssignmentForPosition(@Param("posId")Integer posId);
+	public Assignment getPrimaryAssignmentForPosition(@Param("posId")Long posId);
 
 }
