+++ source/egov-erp/egov-eis/src/main/java/org/egov/eis/repository/PositionHierarchyRepository.java
@@ -17,9 +17,9 @@ import org.springframework.stereotype.Repository;
 public interface PositionHierarchyRepository extends JpaRepository<PositionHierarchy, Integer>{
 
     @Query(" from PositionHierarchy P where P.fromPosition.id=:fromPosition and P.objectType.id=:objectType")
-    PositionHierarchy getPositionHierarchyByPosAndObjectType(@Param("fromPosition")Integer fromPosition,@Param("objectType")Integer objectType);
+    PositionHierarchy getPositionHierarchyByPosAndObjectType(@Param("fromPosition")Long fromPosition,@Param("objectType")Integer objectType);
     
     @Query(" from PositionHierarchy P where P.fromPosition.id=:fromPosition and P.objectType.id=:objectType and P.objectSubType=:objectSubType")
-    PositionHierarchy getPosHirByPosAndObjectTypeAndObjectSubType(@Param("fromPosition")Integer fromPosition,@Param("objectType")Integer objectType,@Param("objectSubType")String objectSubType);
+    PositionHierarchy getPosHirByPosAndObjectTypeAndObjectSubType(@Param("fromPosition")Long fromPosition,@Param("objectType")Integer objectType,@Param("objectSubType")String objectSubType);
     
 }
