+++ source/egov-erp/egov-eis/src/main/java/org/egov/eis/service/AssignmentService.java
@@ -39,11 +39,11 @@ public class AssignmentService {
         return assignmentRepository.getAllActiveAssignmentsByEmpId(Id);
     }
     
-    public Assignment getAssignmentForPosition(final Integer posId,final Date givenDate) {
+    public Assignment getAssignmentForPosition(final Long posId,final Date givenDate) {
         return assignmentRepository.getAssignmentsForPosition(posId,givenDate).get(0);
     }
     
-    public Assignment getPrimaryAssignmentForPositon(final Integer posId) {
+    public Assignment getPrimaryAssignmentForPositon(final Long posId) {
         return assignmentRepository.getPrimaryAssignmentForPosition(posId);
     }
